TIL-Analyzer-main/requirements.txt
autograd==1.8.0
autograd-gamma==0.5.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.13.1
fonttools==4.59.2
formulaic==1.2.0
fsspec==2024.6.1
interface-meta==1.3.0
Jinja2==3.1.3
joblib==1.5.2
kiwisolver==1.4.9
lifelines==0.30.0
llvmlite==0.44.0
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
narwhals==2.4.0
networkx==3.3
numba==0.61.2
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
wrapt==1.17.3
